refactor: cppize llvm_jit_cpu.h
