+++ taichi/backends/llvm_jit_ptx.cpp
@@ -17,6 +17,11 @@ TLANG_NAMESPACE_BEGIN
 
 #if defined(TI_WITH_CUDA)
 
+std::string compile_module_to_ptx(std::unique_ptr<llvm::Module> &module);
+int compile_ptx_and_launch(const std::string &ptx,
+                           const std::string &kernel_name,
+                           void *);
+
 std::string cuda_mattrs() {
   return "+ptx50";
 }
