+++ taichi/llvm_context.h
@@ -14,8 +14,6 @@
 TLANG_NAMESPACE_BEGIN
 class JITSessionCPU;
 
-void *jit_lookup_name(JITSession *jit, const std::string &name);
-
 class TaichiLLVMContext {
  public:
   std::unique_ptr<llvm::LLVMContext> ctx;
@@ -39,11 +37,8 @@ class TaichiLLVMContext {
 
   void add_module(std::unique_ptr<llvm::Module> module);
 
-  llvm::JITSymbol lookup_symbol(const std::string &name);
-
   virtual void *lookup_function_pointer(const std::string &name) {
-    auto func_ptr = jit_lookup_name(jit.get(), name);
-    return func_ptr;
+    return jit->lookup(name);
   }
 
   // Unfortunately, this can't be virtual since it's a template function
